+++ image/nginx-passenger.sh
@@ -10,17 +10,17 @@ header "Installing Phusion Passenger..."
 ## from APT and Ruby installed from RVM.
 if [[ ! -e /usr/bin/ruby ]]; then
 	run /pd_build/ruby_support/prepare.sh
-	run /usr/local/rvm/bin/rvm install ruby-2.3.7
+	run /usr/local/rvm/bin/rvm install ruby-2.5.3
 	# Make passenger_system_ruby work.
-	run create_rvm_wrapper_script ruby2.3 ruby-2.3.7 ruby
+	run create_rvm_wrapper_script ruby2.5 ruby-2.5.3 ruby
 	run /pd_build/ruby_support/finalize.sh
 fi
 
 ## Install Phusion Passenger.
 if [[ "$PASSENGER_ENTERPRISE" ]]; then
-	run apt-get install -y nginx-extras passenger-enterprise
+	run apt-get install -y nginx passenger-enterprise libnginx-mod-http-passenger-enterprise
 else
-	run apt-get install -y nginx-extras passenger
+	run apt-get install -y nginx passenger libnginx-mod-http-passenger
 fi
 run cp /pd_build/config/30_presetup_nginx.sh /etc/my_init.d/
 run cp /pd_build/config/nginx.conf /etc/nginx/nginx.conf
@@ -49,18 +49,6 @@ fi
 if [[ -e /usr/bin/ruby2.3 ]]; then
 	run ruby2.3 -S passenger-config build-native-support
 	run setuser app ruby2.3 -S passenger-config build-native-support
-fi
-if [[ -e /usr/bin/ruby2.2 ]]; then
-	run ruby2.2 -S passenger-config build-native-support
-	run setuser app ruby2.2 -S passenger-config build-native-support
-fi
-if [[ -e /usr/bin/ruby2.1 ]]; then
-	run ruby2.1 -S passenger-config build-native-support
-	run setuser app ruby2.1 -S passenger-config build-native-support
-fi
-if [[ -e /usr/bin/ruby2.0 ]]; then
-	run ruby2.0 -S passenger-config build-native-support
-	run setuser app ruby2.0 -S passenger-config build-native-support
 fi
 if [[ -e /usr/bin/jruby ]]; then
 	run jruby --dev -S passenger-config build-native-support