+++ image/ruby-2.5.3.sh
@@ -2,13 +2,13 @@
 set -e
 source /pd_build/buildconfig
 
-RVM_ID=`basename "$0" | sed 's/.sh$//'`
+RVM_ID=$(basename "$0" | sed 's/.sh$//')
 
 header "Installing $RVM_ID"
 run /pd_build/ruby_support/prepare.sh
 run /usr/local/rvm/bin/rvm install $RVM_ID
 run /usr/local/rvm/bin/rvm-exec $RVM_ID@global gem install $DEFAULT_RUBY_GEMS --no-document
 # Make passenger_system_ruby work.
-run create_rvm_wrapper_script ruby2.4 $RVM_ID ruby
+run create_rvm_wrapper_script ruby2.5 $RVM_ID ruby
 run /pd_build/ruby_support/install_ruby_utils.sh
 run /pd_build/ruby_support/finalize.sh