+++ image/ruby_support/install_ruby_utils.sh
@@ -5,7 +5,6 @@ source /pd_build/buildconfig
 ## The Rails asset compiler requires a Javascript runtime.
 if [[ ! -e /usr/bin/node ]]; then
 	run minimal_apt_get_install nodejs
-	run ln -s /usr/bin/nodejs /usr/bin/node
 fi
 
 ## Install development headers for native libraries that tend to be used often by Ruby gems.
@@ -18,8 +17,6 @@ if ! [[ -e /tmp/ruby_native_libs_installed ]]; then
 	run minimal_apt_get_install libsqlite3-dev
 	## For postgres and pg.
 	run minimal_apt_get_install libpq-dev
-	## For curb.
-	run minimal_apt_get_install libcurl4-openssl-dev
 	## For all kinds of stuff.
 	run minimal_apt_get_install zlib1g-dev